Tutorial.Lib

Injection/Projection pairs

class Injection a b where

Methods

inject :: a -> b

class Injection b a => Projection a b where

Methods

project :: a -> Maybe b

prop_Projection1 :: forall b a. (Arbitrary a, Eq a, Projection a b) => a -> Bool

prop_Projection2 :: forall b a. (Arbitrary a, Eq a, Projection b a) => a -> Bool

Finite maps from Fin n

type Map a n = Fin n -> a

empty :: Map a Z

the empty context has an empty domain

extend :: Map a n -> a -> Map a (S n)

extend a map with a new definition, enlarging its domain

Display class for pretty printing with associated state, such as prececence

class Display s d where

Methods

display :: d -> s -> Doc e

multiple :: Display s a => String -> [a] -> s -> Doc e

Display the list of elements in parentheses, punctuated by the given string. (If there is only one element, don't include the parentheses)

displayMaybe :: (t -> s -> Doc d) -> Maybe t -> s -> Doc d

displayEither :: (Display s a, Display s b) => (forall a. Display s a => a -> s -> Doc d) -> Either a b -> s -> Doc d