Tutorial.Named.PP

Description

A module to display named syntax terms nicely.

This pretty printer is built using the Prettyprinter library, which includes a number of definitions for turning data into Doc

Documentation

oneline :: Display DispState a => a -> String

Pretty-print on a single line (using group to collapse line breaks)

pp :: Display DispState a => a -> String

Pretty-print with possible line breaks

Pretty printing parameters and state

data DispState

Parameters and state for pretty printing

Constructors

DI

Fields

style :: Style

printing style

prec :: Int

precedence level

data Style

Printing style controls how type names are displayed

Constructors

Math
Use 0, 1, 2 for type names

Lang
Use Void, Unit, Bool for type names

initState :: DispState

parens :: Int -> Doc d -> DispState -> Doc d

If the precedence of the current form (level) is less than the precedence of the context (prec s) add parentheses.

levelBase :: Int

Most syntactic forms have precedence level 0

levelBranch :: Int

levelApp :: Int

withPrec :: MonadReader DispState m => Int -> m a -> m a

display0 :: Display DispState a => a -> DispState -> Doc e

test :: Display DispState a => a -> Doc e

For testing: run the pretty printer

Display instance for types

dunit :: DispState -> Doc e

display the unit type

dvoid :: DispState -> Doc e

display the void type

dbool :: DispState -> Doc e

display the bool type

Diplay instance for Terms

displayBr :: (Tm, Tm) -> DispState -> Doc e

Pretty-print a single case branch p -> e

displayBrs :: [(Tm, Tm)] -> DispState -> [Doc e]

Pretty-print a list of case branches, adding | separators when there are multiple

gatherBinders :: Tm -> ([String], Tm)

Collect consecutive lambda binders so they can be printed as λ x y z. e

example :: Tm