Tutorial.Named.Parser

Description

A parsec-based parser for the concrete syntax of the tutorial language, producing Tm and Ty values.

The concrete syntax supports:

Types: Unit/1, Void/0, Bool/2, t -> t, t * t, t + t

Terms: variables, lambda, application, pairs, injections, pattern matching, let, if/then/else, type annotations

Patterns: variables, unit, tuples, injections, annotated patterns

Top-level

type Parser a = Parsec String () a

testParser :: Parser t -> String -> Either ParseError t

Test a Parser on a String

parseTm :: String -> Either ParseError Tm

Parse an expression

Lexer

tokenizer :: TokenParser u

identifier :: Parser String

symbol :: String -> Parser String

whiteSpace :: Parser ()

lexeme :: Parser a -> Parser a

dot :: Parser ()

comma :: Parser ()

parens :: Parser a -> Parser a

brackets :: Parser a -> Parser a

braces :: Parser a -> Parser a

reserved :: String -> Parser ()

reservedOp :: String -> Parser ()

natural :: Parser Int

Types

voidP :: Parser Ty

unitP :: Parser Ty

boolP :: Parser Ty

prodOrSumTy :: Parser Ty -> Parser Ty

Parse a sequence of types separated by either * or +. Once the operator is seen the rest of the list must use the same operator. Factored to avoid backtracking after the first element.

ty :: Parser Ty

toplevel parser for types

Patterns

var :: Parser Tm

inj :: Parser Tm -> Parser Tm

Parse an injection: Inj i e, True, or False

tuple :: Parser Tm -> Parser Tm

Parse a parenthesised tuple: (), (e), or (e1, ..., en)

multiple :: Parser Tm -> Parser Tm

Parse a term that may be a product or a single term. Both forms begin with a single term; we commit after seeing ,.

pat :: Parser Tm

A pattern produces an element of the term datatype but in a restricted set

lambda :: Parser Tm

Parse a lambda abstraction: x1 ... xn . e or λ x1 ... xn . e

casetm :: Parser Tm

Parse a case expression: case e of | p -> e | ...

branch :: Parser (Tm, Tm)

Parse a single branch: p -> e

seq :: Parser Tm

sequence

tm :: Parser Tm

Parse a term (entry point for expressions)

lettm :: Parser Tm

Parse a let expression: let x = e1 in e2 (sugar for ( x . e2) e1)

iftm :: Parser Tm

Parse an if expression: if e then e1 else e2 (sugar for case)